+++ test/onnx/parse/if_then_test.cpp
@@ -40,11 +40,11 @@ TEST_CASE(if_then_test)
     auto y    = mm->add_parameter("y", s);
     auto cond = mm->add_parameter("cond", sc);
 
-    auto* then_mod = p.create_module("If_5_if");
+    auto* then_mod = p.create_module("main_If_5_if");
     auto rt        = then_mod->add_instruction(migraphx::make_op("add"), x, l1);
     then_mod->add_return({rt});
 
-    auto* else_mod = p.create_module("If_5_else");
+    auto* else_mod = p.create_module("main_If_5_else");
     auto re        = else_mod->add_instruction(migraphx::make_op("mul"), y, l2);
     else_mod->add_return({re});